chore(rename) mod Agreement -> Dkg
