+++ src/lib.rs
@@ -87,8 +87,8 @@ pub use xor_name::{Prefix, XorName, XOR_NAME_LEN}; // TODO remove pub on API upd
 // Private
 // ############################################################################
 
-mod agreement;
 mod cache;
+mod dkg;
 mod ed25519;
 mod error;
 mod event;
