+++ src/routing/split_barrier.rs
@@ -8,7 +8,7 @@
 
 use std::mem;
 
-use crate::agreement::Signed;
+use crate::dkg::Signed;
 use sn_messaging::{node::Proven, SectionAuthorityProvider};
 use xor_name::Prefix;
